Reorganize file structure by moving components and composables to dedicated folders
